+++ packages/plugin-asset/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @stellar-agent-kit/plugin-asset
 
+## 0.1.8
+
+### Patch Changes
+
+- @stellar-agent-kit/core@0.1.8
+
 ## 0.1.7
 
 ### Patch Changes